+++ internal/pkg/util/shell/interpreter/interpreter.go
@@ -147,7 +147,8 @@ func New(r io.Reader, name string, args []string, envs []string, runnerOptions .
 	}
 
 	// TODO - update to ExecHandlers, as ExecHandler is deprecated.
-	opts := []interp.RunnerOption{
+	opts := make([]interp.RunnerOption, 0, 7+len(runnerOptions))
+	opts = append(opts,
 		interp.StdIO(os.Stdin, os.Stdout, os.Stderr),
 		interp.ExecHandler(s.internalExecHandler()), //nolint:staticcheck
 		interp.OpenHandler(s.internalOpenHandler()),
@@ -155,7 +156,7 @@ func New(r io.Reader, name string, args []string, envs []string, runnerOptions .
 		interp.Params("--"),
 		interp.Env(expand.ListEnviron(envs...)),
 		interp.Dir(dir),
-	}
+	)
 	opts = append(opts, runnerOptions...)
 	s.runner, err = interp.New(opts...)
 	if err != nil {